Move constants/types/utils to separate files
