+++ ui/src/dashboards/constants/index.ts
@@ -118,3 +118,6 @@ export enum CEOTabs {
 export const MAX_TOLOCALESTRING_VAL = 20 // 20 is the max input to maximumFractionDigits in spec for tolocalestring
 export const MIN_DECIMAL_PLACES = '0'
 export const MAX_DECIMAL_PLACES = MAX_TOLOCALESTRING_VAL.toString()
+
+// used in importing dashboards and mapping sources
+export const NO_SOURCE = 'none'
